# Cached sequential access of GetPhysicalPointer