Often memory is accessed in sequence, checking the
previous area before searching other areas reduces
the cost of lookup.

Tested running Pokemon Moon with
Visual Studio 2017 Profiler (100,000~ samples):

BaseLine:
Memory::GetPhysicalPointer    1.8% CPU samples (exclusive)
SVC::CallSVC                  56.71% CPU samples (inclusive)

With AreaCache:
Memory::GetPhysicalPointer    0.4% CPU samples (exclusive)
SVC::CallSVC                  55.01% CPU samples (inclusive)

<!-- Reviewable:start -->
---
This change is [<img src="https://reviewable.io/review_button.svg" height="34" align="absmiddle" alt="Reviewable"/>](https://reviewable.io/reviews/citra-emu/citra/3026)
<!-- Reviewable:end -->