@yuriks 2017-10-21
Also forgot to mention: Add a comment explaining the what and why of the optimization.

Echoing some questions from IRC: What effect does this have on frametimes in games, etc.?