@acnleditor2 2017-10-23
some games cause citra to crash when starting / after loading