@acnleditor2 2017-10-26
pokemon sun & moon crash on opening qr reader