@Subv 2017-12-14
I don't like the idea of adding _yet another cache and static state_ to the emulator for something that will probably not benefit speed much if at all in the grand scheme of things. This imo falls in the realm of microoptimization, there are bigger targets out there.