@B3n30 2017-12-16
Closed because there is reportedly no improvements out of this PR and it just adds complexity to the code