@tabnk 2017-12-29
Notice Super Mario 3D Land slower after this PR was removed from canary build so I’m decided test some Mario games.

Master Build on Haswell CPU
Super Mario 3D = 64 fps 
Mario Kart - 68 fps 

This PR on Haswell CPU
Super Mario 3D = 67 fps 
Mario Kart - 72 fps 

Overall still a nice performance improvement.